feat: implement expense-service-refactor spec - extract 6 sub-services from monolithic expenseService.js
