--- .kiro/hooks/no-console-backend.kiro.hook
@@ -1,23 +1,0 @@
-{
-  "enabled": true,
-  "name": "No Console in Backend",
-  "description": "When backend source files (not tests or scripts) are edited, checks for console.log/error/warn usage and reminds to use the logger module instead.",
-  "version": "1",
-  "when": {
-    "type": "fileEdited",
-    "patterns": [
-      "backend/controllers/*.js",
-      "backend/services/*.js",
-      "backend/repositories/*.js",
-      "backend/routes/*.js",
-      "backend/middleware/*.js",
-      "backend/database/*.js",
-      "backend/utils/*.js",
-      "backend/server.js"
-    ]
-  },
-  "then": {
-    "type": "askAgent",
-    "prompt": "Check the edited file for any console.log, console.error, or console.warn statements. If found, replace them with the appropriate logger method (logger.debug, logger.info, logger.warn, logger.error) from require('../config/logger') or require('../../config/logger'). Exception: files in backend/scripts/ and test files are allowed to use console statements."
-  }
-}
