--- .kiro/specs/shared-data-context/.config.kiro
@@ -1,1 +1,0 @@
-{"generationMode": "requirements-first"}
